feat: 🐊 added types for defining rules, added hyper key helper method, defined hyper key rule
just wanted to get something started here, not sure if I will end up with this data model. Wanted to make it
so you could just write code to define the config and remove the need for json as much as possible.

I think this update gets close to that, will now incrementally improve it.